src/clpp11.hpp: use const-ness coherently
Settle on a semantics of const-ness in data types wrapping OpenCL objects.
In particular, replacing a pointed-to object is non-const. Any other action
is const because it can be achieved via direct C API call on the internal
object retrieved via operator()() const.

This matches the semantics of a (shared) pointer.